Rename manifest to workspace
